Square Wave
Got a square wave revenue generator working.
The value investors now use the exponentially smoothed weighted average of past revenue (might want to check that the time discount is applied properly).
The economy doesn't get to a good equilibrium because the value investors buy in the beginning and then the trend and regression investors buy eventually.  Then when the revenue drops, the value investors don't have enough shares to sell to make the price go to their spot price.  The trendline investors starts selling, but the reversion to mean investor starts buying, buoying the price up above true value.